feat(version): 将接口版本号从整型改为小数支持精确控制
- 接口版本字段从 int 类型改为 BigDecimal 类型，支持小数版本控制
- 数据库 schema 中 version 字段从 INT 改为 DECIMAL(10,1)，默认值从 1 改为 1.0
- 版本递增策略从 +1 改为 +0.1，实现更精细的版本管理
- 更新所有相关实体类、DTO、Repository 和 Service 层的版本字段类型
- 修改测试用例以适配新的小数版本格式
- 调整前端界面显示逻辑，将版本计算从 +1 改为 +0.1 并保留一位小数
- 修复集成测试中的版本比较逻辑，使用 BigDecimal 的比较方法
- 在控制器和异常处理中统一使用 BigDecimal 类型的版本参数

+++ src/main/java/com/deepx/apicenter/controller/admin/InterfaceController.java
@@ -36,6 +36,7 @@ import tools.jackson.databind.node.ArrayNode;
 import tools.jackson.databind.node.ObjectNode;
 
 import java.nio.charset.StandardCharsets;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -237,7 +238,7 @@ public class InterfaceController {
 
     /** 快照详情（config_json 完整可重建） */
     @GetMapping("/{id}/versions/{version}")
-    public ApiResult<InterfaceService.VersionDetail> versionDetail(@PathVariable long id, @PathVariable int version) {
+    public ApiResult<InterfaceService.VersionDetail> versionDetail(@PathVariable long id, @PathVariable BigDecimal version) {
         return ApiResult.ok(interfaceService.versionDetail(id, version));
     }
